backend/requirements.txt
flask>=3.0
flask-cors>=4.0
gunicorn>=21.2
transformers>=4.30,<5.0
torch>=2.0
newspaper3k
lxml_html_clean
beautifulsoup4>=4.12
requests>=2.31
